requirements.txt
faster-whisper==1.0.0
webrtcvad==2.0.10
numpy>=1.26
soundfile>=0.12
librosa>=0.10
pydub>=0.25
pandas>=2.2
pyyaml>=6.0
tqdm>=4.66
